Recruit DTO, Entity 분리
추후 데이터를 앱에서 사용하기 편한 형태로 가공하기 위함

Separates the Recruit model into DTO and Entity to improve data handling.

The Recruit struct is renamed to RecruitResponseDTO for representing the
network response. A new RecruitEntity struct is introduced to represent
the data model within the application.

This change allows for better separation of concerns between network
data and the application's data model. It also includes a `toEntity()`
method to map the DTO to the Entity.

The NetworkService is updated to use generics, and RecruitModel and RecruitModelStub are updated to use RecruitEntity.