테스트 yml 추가
Sets up a new CI workflow using GitHub Actions to run tests on develop and main branches, as well as pull requests.

The workflow includes steps to:
- Checkout the code
- Select Xcode 16.0
- Create a Secrets.xcconfig file
- Cache SPM dependencies and DerivedData
- Run tests for iOS 18.0 on iPhone 16 Pro

This ensures that tests are automatically run on each push and pull request, helping to catch errors early and improve the quality of the codebase.